update to stable channel. move benchmark to criterion
